% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot2scales.R
\name{scale_colour_poke}
\alias{scale_color_poke}
\alias{scale_colour_poke}
\alias{scale_fill_poke}
\title{Add a pokemon palette to a ggplot2 colour or fill scale.}
\usage{
scale_colour_poke(..., pokemon = 1)

scale_fill_poke(..., pokemon = 1)

scale_color_poke(..., pokemon = 1)
}
\arguments{
\item{pokemon}{An integer or character pokemon name}
}
\description{
Get a pokemon palette by either giving a pokemon number or name.
}
\examples{
qplot(Sepal.Length, Sepal.Width, colour = Species, data=iris) +
 scale_colour_poke('Metapod')
}

