% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team.R
\name{team}
\alias{team}
\title{Combine colours from different pokemon.}
\usage{
team(pokemon, n = 1, spread = NULL)
}
\arguments{
\item{pokemon}{A vector of integers or a character vector of pokemon name}

\item{n}{How many colours per pokemon?}

\item{spread}{How many, fairly distinct, colours should be returned for each pokemon. 
If not null, \code{n} is ignored.}
}
\description{
Combine colours from different pokemon.
}
