% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/palettetown.R
\name{pokepal}
\alias{pokepal}
\title{Get a pokemon palette.}
\usage{
pokepal(pokemon = 1)
}
\arguments{
\item{pokemon}{An integer or character pokemon name}
}
\description{
Get a pokemon palette by either giving a pokemon number or name.
}
\examples{
pal <- pokepal(3)
plot(1:length(pal), col = pal)
}

